Fix clang warning: -Wimplicit-int-conversion
Most of these warnings are harmless, and are only there because of c/c++
integer promotion rules, but the value can not actually get out of
range.

In a few places I added extra input-string validation (for stuff like
"button1000" but we 'only' support 256 joystick buttons).

After this the clang build is warning free again. Except for some
warnings in the the src/3rdpart subdirectory. Ideally we only compile
3rdparty files with standard warning settings, or without any warnings.
But unfortunately that's not supported in our homegrown build system.

+++ src/fdc/MSXtar.cc
@@ -959,7 +959,7 @@ std::string MSXtar::dir()
 	for (unsigned i = 0; i < num; ++i) {
 		auto entry = list.getListIndexUnchecked(i);
 		auto filename = std::string(entry.getListIndexUnchecked(0).getString());
-		auto attrib = DiskImageUtils::formatAttrib(entry.getListIndexUnchecked(1).getOptionalInt().value_or(0));
+		auto attrib = DiskImageUtils::formatAttrib(narrow<uint8_t>(entry.getListIndexUnchecked(1).getOptionalInt().value_or(0)));
 		//time_t time = entry.getListIndexUnchecked(2).getOptionalInt().value_or(0); // ignored
 		auto size = entry.getListIndexUnchecked(3).getOptionalInt().value_or(0);
