+++ src/memory/RomNeo8.cc
@@ -26,9 +26,9 @@ void RomNeo8::writeMem(word address, byte value, EmuTime::param /*time*/)
 	if (bbb < 2) return;
 	unsigned region = bbb - 2;
 	if ((address & 1) == 0) {
-		blockReg[region] = (blockReg[region] & 0xFF00) | value;
+		blockReg[region] = uint16_t((blockReg[region] & 0xFF00) | value);
 	} else {
-		blockReg[region] = (blockReg[region] & 0x00FF) | ((value & 0b1111) << 8);
+		blockReg[region] = uint16_t((blockReg[region] & 0x00FF) | ((value & 0b1111) << 8));
 	}
 	setRom(region, blockReg[region]);
 }
